pyproject.toml
[project]
name = "aoc25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mypy>=1.19.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "ruff>=0.14.8",
]
